+++ src/components/pages/pdf/product-information.js
@@ -195,9 +195,9 @@ export const ProductInformation = () => (
             <div>
               In minutes. Visit our{' '}
               <Link href='/docs/guides/pdf'>documentation</Link> for interactive
-              playground examples, official{' '}
-              <Link href='/integrations/sdk'>SDKs</Link> (Node.js, Python, Ruby,
-              Go), and copy-paste code snippets for any language.
+              playground examples, the official{' '}
+              <Link href='/integrations/sdk'>SDK</Link> (Node.js, browsers,
+              Deno), and copy-paste code snippets for any language.
             </div>
             <div>
               Need architectural advice or have custom requirements? Check our{' '}